LayoutTests/js/typedarray-some-expected.txt

This test checks the behavior of the TypedArray.prototype.some function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Int32Array.prototype.some.length is 1
PASS Int32Array.prototype.some.name is 'some'
PASS isSameFunctionForEachTypedArrayPrototype('some') is true
PASS testPrototypeReceivesArray('some', [undefined, this, { }, [ ], true, '']) is true

testPrototypeFunction has the following arg list (name, args, init, result [ , expectedArray ])

1.0 Single Argument Testing
PASS testPrototypeFunction('some', '(isBigEnough)', [12, 5, 8, 13, 44], true, [12, 5, 8, 13, 44]) is true
PASS testPrototypeFunction('some', '(isBigEnough)', [2, 4, 8, 3, 4], false) is true

2.0 Two Argument Testing
PASS testPrototypeFunction('some', '(isBigEnough, thisValue)', [2, 5, 10, 3, 4], false) is true
PASS testPrototypeFunction('some', '(isBigEnough, thisValue)', [12, 54, 82, 13, 44], true) is true

3.0 Array Element Changing
PASS testPrototypeFunction('some', '(isBigEnoughAndChange)', [2, 5, 1, 13, 44], false, [5, 5, 5, 5, 5]) is true
PASS testPrototypeFunction('some', '(isBigEnoughAndChange)', [12, 15, 10, 13, 44], true, [12, 15, 10, 13, 5]) is true

4.0 Exception Test
PASS testPrototypeFunction('some', '(isBigEnoughAndException)', [12, 15, 10, 13, 44], true) is true
PASS testPrototypeFunction('some', '(isBigEnoughAndException)', [1, 15, 10, 13, 44], false) threw exception exception from function.

5.0 Wrong Type for Callback Test
PASS testPrototypeFunction('some', '(8)', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.some callback must be a function.
PASS testPrototypeFunction('some', '("wrong")', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.some callback must be a function.
PASS testPrototypeFunction('some', '(new Object())', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.some callback must be a function.
PASS testPrototypeFunction('some', '(null)', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.some callback must be a function.
PASS testPrototypeFunction('some', '()', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.some callback must be a function.

PASS successfullyParsed is true

TEST COMPLETE
